requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coinbase-commerce==1.0.1
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
daphne==4.0.0
Django==4.2.2
django-cors-headers==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.5.0
drf-yasg==1.21.6
filetype==1.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
msgpack==1.0.5
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.3
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
zope.interface==6.0

